fix: resolve OrderListItemDto type mismatch and unify order service
