+++ app/_module/api/types.ts
@@ -63,44 +63,10 @@ export type CreateOrderResponseDto = Schemas['CreateOrderResponseDto'];
 export type OrderAttendeeDto = Schemas['OrderAttendeeDto'];
 export type OrderGifterDto = Schemas['OrderGifterDto'];
 export type OrderedTicketDto = Schemas['OrderedTicketDto'];
-// export type OrderListItemDto = Schemas['OrderListItemDto'];
+export type OrderListItemDto = Schemas['OrderListItemDto'];
 export type OrdersPaginationMetaDto = Schemas['OrdersPaginationMetaDto'];
 export type OrderListResponseDto = Schemas['OrderListResponseDto'];
-// export type GetOrderReferenceResponseDto = Schemas['GetOrderReferenceResponseDto'];
-
-/** GET /api/v1/orders/reference/{reference} — 200 response */
-export interface GetOrderReferenceResponseDto {
-  ticket: {
-    name: string;
-    /** Signed URL to the ticket PDF */
-    url: string;
-    validityDates: string[];
-  };
-  /** Amount paid in Naira (2 decimal places), e.g. "9500.00" */
-  amount: string;
-  status:
-    'AWAITING_PAYMENT' | 'PAID' | 'CANCELLED' | 'AWAITING_REFUND' | 'REFUNDED';
-  /** Ticket code used as the download token */
-  code: string;
-}
-
-/** Single item from GET /api/v1/orders list */
-export interface OrderListItemDto {
-  id: string;
-  paidAt: string | null;
-  amount: string;
-  status:
-    'AWAITING_PAYMENT' | 'PAID' | 'CANCELLED' | 'AWAITING_REFUND' | 'REFUNDED';
-  attendeeFullName: string;
-  attendeeEmail: string;
-  checkIns: string[];
-  ticket?: {
-    id: string;
-    name: string;
-    code: string;
-    validity: string;
-  };
-}
+export type GetOrderReferenceResponseDto = Schemas['GetOrderReferenceResponseDto'];
 
 // ---- Query param helpers -----------------------------------
 export interface OrderListParams {
